+++ glutin/src/platform/emscripten/mod.rs
@@ -72,11 +72,11 @@ impl Context {
         el: &winit::EventsLoop,
         pf_reqs: &PixelFormatRequirements,
         gl_attr: &GlAttributes<&Context>,
-        dims: dpi::PhysicalSize,
+        size: dpi::PhysicalSize,
     ) -> Result<Self, CreationError> {
         let wb = winit::WindowBuilder::new()
             .with_visibility(false)
-            .with_dimensions(dims.to_logical(1.));
+            .with_dimensions(size.to_logical(1.));
 
         Self::new_windowed(wb, el, pf_reqs, gl_attr).map(|(w, c)| match c {
             Context::Window(c) => Context::WindowedContext(w, c),