+++ glutin_examples/build.rs
@@ -9,7 +9,7 @@ fn main() {
     println!("cargo:rerun-if-changed=build.rs");
 
     let mut file = File::create(&dest.join("gl_bindings.rs")).unwrap();
-    Registry::new(Api::Gles2, (3, 0), Profile::Core, Fallbacks::All, [])
+    Registry::new(Api::Gles2, (3, 3), Profile::Core, Fallbacks::All, [])
         .write_bindings(gl_generator::StructGenerator, &mut file)
         .unwrap();
 }